+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 **New Features and Improvements**
 
 * Runtime: added a new Mover component for controlling characters that implements a standard collide-and-slide algorithm.
+* Runtime: added a new Global Lighting component for controlling the skydome's appearance and other global lighting parameters.
 * Runtime: added a new dedicated skydome renderer.
 * Runtime: added a new .render_config resource for configuring the renderer, along with a `render_config` property in the boot files to allow selecting the desired config.
 * Runtime: added a new `spinning_friction` property to physics materials that you can specify in `global.physics_config`.